更新

+++ src/main/java/cms/web/action/topic/TagManageAction.java
@@ -1,6 +1,8 @@
 package cms.web.action.topic;
 
 
+import java.io.File;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -14,7 +16,11 @@ import cms.bean.ResultCode;
 import cms.bean.topic.Tag;
 import cms.service.setting.SettingService;
 import cms.service.topic.TagService;
+import cms.utils.FileUtil;
 import cms.utils.JsonUtils;
+import cms.utils.UUIDUtil;
+import cms.web.action.TextFilterManage;
+import cms.web.action.fileSystem.FileManage;
 
 import org.springframework.context.MessageSource;
 import org.springframework.stereotype.Controller;
@@ -25,9 +31,10 @@ import org.springframework.validation.Validator;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.multipart.MultipartFile;
 
 /**
- * 标签
+ * 标签管理
  *
  */
 @Controller
@@ -39,9 +46,10 @@ public class TagManageAction {
 	@Resource TagService tagService; 
 	@Resource TagManage tagManage;
 	
-	
+	@Resource FileManage fileManage;
 	@Resource SettingService settingService;
 	@Resource MessageSource messageSource;
+	@Resource TextFilterManage textFilterManage;
 	
 	/**
 	 * 标签   添加界面显示
@@ -59,10 +67,83 @@ public class TagManageAction {
 	 */
 	@RequestMapping(params="method=add", method=RequestMethod.POST)
 	@ResponseBody
-	public String add(ModelMap model,Tag formbean,BindingResult result
-			) throws Exception {
+	public String add(ModelMap model,Tag formbean,BindingResult result,String imagePath,MultipartFile images,
+			HttpServletRequest request, HttpServletResponse response) throws Exception {
 		//错误
 		Map<String,Object> error = new HashMap<String,Object>();
+		Tag tag = new Tag(); 
+		String _imagePath = "";
+		String _fileName = "";
+		if (images == null || images.isEmpty()) { 
+			if(imagePath != null && !"".equals(imagePath.trim())){
+				imagePath = textFilterManage.deleteBindURL(request, imagePath);
+				
+				String fileName = FileUtil.getName(imagePath);
+				
+				//取得路径名称
+				String pathName = FileUtil.getFullPath(imagePath);
+				
+				//旧路径必须为file/topicTag/开头
+				if(imagePath.substring(0, 14).equals("file/topicTag/")){
+					//新路径名称
+					String newPathName = "file/topicTag/";
+					
+					//如果新旧路径不一致
+					if(!newPathName.equals(pathName)){
+						
+						//复制文件到新路径
+						fileManage.copyFile(FileUtil.toRelativePath(imagePath), newPathName);
+						//新建文件锁到新路径
+						//生成锁文件名称
+						String lockFileName =fileName;
+						//添加文件锁
+						fileManage.addLock("file"+File.separator+"topicTag"+File.separator+"lock"+File.separator,lockFileName);
+						
+						
+					}
+				
+					_imagePath = "file/topicTag/";
+					_fileName = fileName;
+				}
+				
+					  
+			}
+        }else{
+        	//验证文件类型
+			List<String> formatList = new ArrayList<String>();
+			formatList.add("gif");
+			formatList.add("jpg");
+			formatList.add("jpeg");
+			formatList.add("bmp");
+			formatList.add("png");
+			formatList.add("svg");
+			boolean authentication = FileUtil.validateFileSuffix(images.getOriginalFilename(),formatList);
+			if(authentication){
+				//取得文件后缀		
+				String ext = FileUtil.getExtension(images.getOriginalFilename());
+				//文件保存目录;分多目录主要是为了分散图片目录,提高检索速度
+				String pathDir = "file"+File.separator+"topicTag"+File.separator;
+				//构建文件名称
+				String fileName = UUIDUtil.getUUID32()+ "." + ext;
+				_imagePath = "file/topicTag/";
+				_fileName = fileName; 
+				  
+				//生成文件保存目录
+				FileUtil.createFolder(pathDir);
+				 
+				//生成锁文件名称
+				String lockFileName = fileName;
+				//添加文件锁
+				fileManage.addLock("file"+File.separator+"topicTag"+File.separator+"lock"+File.separator,lockFileName);
+				
+				//保存文件
+				fileManage.writeFile(pathDir, fileName,images.getBytes());
+				
+			}else{
+				error.put("images", "图片格式错误");
+			}
+        }
+		tag.setImage(_imagePath+_fileName);
 		
 		//数据校验
 		this.validator.validate(formbean, result); 
@@ -75,13 +156,18 @@ public class TagManageAction {
 			}
 		} 
 		if(error.size() ==0){
-			Tag tag = new Tag(); 
+			
 			tag.setId(tagManage.nextNumber());
 			tag.setName(formbean.getName());
 			
 			tag.setSort(formbean.getSort());
 				
 			tagService.saveTag(tag);
+			
+			//删除图片锁
+			if(_imagePath != null && !"".equals(_imagePath.trim())){	
+				fileManage.deleteLock("file"+File.separator+"topicTag"+File.separator+"lock"+File.separator,_fileName);		
+			}
 		}
 		
 		if(error.size() >0){
@@ -101,10 +187,16 @@ public class TagManageAction {
 			HttpServletRequest request, HttpServletResponse response) throws Exception {
 		//错误
 		Map<String,Object> error = new HashMap<String,Object>();
+		Map<String,Object> returnValue = new HashMap<String,Object>();
 		if(tagId != null){//判断ID是否存在;
 			Tag tag = tagService.findById(tagId);
 			if(tag != null){
-				return JsonUtils.toJSONString(new RequestResult(ResultCode.SUCCESS,tag));
+				
+				if(tag.getImage() != null && !"".equals(tag.getImage())){
+					returnValue.put("imagePath",fileManage.fileServerAddress(request)+tag.getImage());
+				}
+				returnValue.put("tag",tag);
+				return JsonUtils.toJSONString(new RequestResult(ResultCode.SUCCESS,returnValue));
 			}else{
 				error.put("tag", "标签不存在");
 			}
@@ -120,16 +212,88 @@ public class TagManageAction {
 	 */
 	@ResponseBody
 	@RequestMapping(params="method=edit", method=RequestMethod.POST)
-	public String edit(ModelMap model,Tag formbean,BindingResult result,Long tagId,
+	public String edit(ModelMap model,Tag formbean,BindingResult result,Long tagId,String imagePath,
+			MultipartFile images,
 			HttpServletRequest request, HttpServletResponse response) throws Exception {
 		//错误
 		Map<String,Object> error = new HashMap<String,Object>();
 		
 		Tag tag = null;
+		String _imagePath = "";
+		String _fileName = "";
+		
 		if(tagId != null && tagId >0L){
 			//取得对象
 			tag = tagService.findById(tagId);
-			if(tag == null){
+			if(tag != null){
+				if(images ==null || images.isEmpty()){//如果图片已上传
+					if(imagePath != null && !"".equals(imagePath.trim())){
+						imagePath = textFilterManage.deleteBindURL(request, imagePath);
+						
+						//取得文件名称
+						String fileName = FileUtil.getName(imagePath);
+
+						//取得路径名称
+						String pathName = FileUtil.getFullPath(imagePath);
+						
+						//旧路径必须为file/topicTag/开头
+						if(imagePath.substring(0, 14).equals("file/topicTag/")){
+							//新路径名称
+							String newPathName = "file/topicTag/";
+							
+							//如果新旧路径不一致
+							if(!newPathName.equals(pathName)){
+								
+								//复制文件到新路径
+								fileManage.copyFile(FileUtil.toRelativePath(imagePath), newPathName);
+								//新建文件锁到新路径
+								//生成锁文件名称
+								String lockFileName = fileName;
+								//添加文件锁
+								fileManage.addLock("file"+File.separator+"topicTag"+File.separator+"lock"+File.separator,lockFileName);
+			
+								
+							}
+							_imagePath = "file/topicTag/";
+							_fileName = fileName;
+						}
+					}
+				}
+				
+				if(images !=null && !images.isEmpty()){	
+					//验证文件类型
+					List<String> formatList = new ArrayList<String>();
+					formatList.add("gif");
+					formatList.add("jpg");
+					formatList.add("jpeg");
+					formatList.add("bmp");
+					formatList.add("png");
+					formatList.add("svg");
+					boolean authentication = FileUtil.validateFileSuffix(images.getOriginalFilename(),formatList);
+					if(authentication){
+						//取得文件后缀		
+						String ext = FileUtil.getExtension(images.getOriginalFilename());
+						//文件保存目录;分多目录主要是为了分散图片目录,提高检索速度
+						String pathDir = "file"+File.separator+"topicTag"+File.separator;
+						//构建文件名称
+						String fileName = UUIDUtil.getUUID32()+ "." + ext;
+						_imagePath = "file/topicTag/";
+						_fileName = fileName;
+						   
+						//生成文件保存目录
+						FileUtil.createFolder(pathDir);
+						//生成锁文件名称
+						String lockFileName = fileName;
+						//添加文件锁
+						fileManage.addLock("file"+File.separator+"topicTag"+File.separator+"lock"+File.separator,lockFileName);
+						  
+						//保存文件
+						fileManage.writeFile(pathDir, fileName,images.getBytes());
+				   }else{
+						error.put("images", "图片格式错误");
+				   }
+				}
+			}else{
 				error.put("tag", "标签不存在");
 			}
 		}else{
@@ -151,10 +315,26 @@ public class TagManageAction {
 		if(error.size() == 0){
 			Tag new_tag = new Tag(); 
 			new_tag.setId(tagId);
+			new_tag.setImage(_imagePath+_fileName);
 			new_tag.setName(formbean.getName());
 			new_tag.setSort(formbean.getSort());
 			
 			tagService.updateTag(new_tag);
+			
+			if(tag.getImage() != null && !"".equals(tag.getImage().trim())){
+				if(!(_imagePath+_fileName).equals(tag.getImage())){//如果图片有变化
+					//删除旧图片
+					//替换路径中的..号
+					String oldPathFile = FileUtil.toRelativePath(tag.getImage());
+					//删除旧文件
+					fileManage.deleteFile(FileUtil.toSystemPath(oldPathFile));
+				}
+			}
+			
+			//删除图片锁
+			if(_imagePath != null && !"".equals(_imagePath.trim())){	
+				fileManage.deleteLock("file"+File.separator+"topicTag"+File.separator+"lock"+File.separator,_fileName);		
+			}
 		}
 		
 		
@@ -180,11 +360,26 @@ public class TagManageAction {
 		//错误
 		Map<String,String> error = new HashMap<String,String>();
 		if(tagId != null && tagId >0L){
-			int i = tagService.deleteTag(tagId);
-			if(i >0){
-				return JsonUtils.toJSONString(new RequestResult(ResultCode.SUCCESS,null));
-			}
+			Tag tag = tagService.findById(tagId);
+			if(tag != null){
+				
 			
+				if(tag.getImage() != null && !"".equals(tag.getImage().trim())){
+					
+					//删除旧图片
+					//替换路径中的..号
+					String oldPathFile = FileUtil.toRelativePath(tag.getImage());
+					//删除旧文件
+					fileManage.deleteFile(FileUtil.toSystemPath(oldPathFile));
+					
+				}
+				int i = tagService.deleteTag(tagId);
+				if(i >0){
+					return JsonUtils.toJSONString(new RequestResult(ResultCode.SUCCESS,null));
+				}
+			}else{
+				error.put("tagId", " 标签不存在");
+			}
 		}else{
 			error.put("tagId", " 标签Id不能为空");
 		}
